pyproject.toml
[project]
name = "workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.84.0",
    "beautifulsoup4>=4.14.3",
    "bertopic>=0.17.4",
    "cpnet>=0.0.21",
    "hdbscan>=0.8.42",
    "httpx>=0.28.1",
    "matplotlib>=3.10.8",
    "networkx>=3.6.1",
    "openai>=2.26.0",
    "pandas>=3.0.1",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "scipy>=1.17.1",
    "sentence-transformers>=5.4.0",
    "umap-learn>=0.5.12",
]
